fix: SCN_DrawSurface BaseVertexLocation=0 (world IB uses absolute indices); fix Z_Malloc size_t cast
Agent-Logs-Url: https://github.com/emmyie/etlegacy/sessions/a9de60b5-f728-452c-819a-f3d3e2cda96e

+++ src/renderer_dx12/dx12_world.cpp
@@ -675,8 +675,8 @@ void DX12_LoadWorld(const char *name)
 		if (count > 0)
 		{
 			// ---- Allocate parallel arrays for batch upload ----
-			ID3D12Resource **texResources    = (ID3D12Resource **)dx12.ri.Z_Malloc(count * (int)sizeof(ID3D12Resource *));
-			ID3D12Resource **uploadBuffers   = (ID3D12Resource **)dx12.ri.Z_Malloc(count * (int)sizeof(ID3D12Resource *));
+			ID3D12Resource **texResources    = (ID3D12Resource **)dx12.ri.Z_Malloc((int)((size_t)count * sizeof(ID3D12Resource *)));
+			ID3D12Resource **uploadBuffers   = (ID3D12Resource **)dx12.ri.Z_Malloc((int)((size_t)count * sizeof(ID3D12Resource *)));
 			HRESULT          hr;
 
 			if (!texResources || !uploadBuffers)
